+++ diffsims/utils/atomic_diffraction_generator_support/discretise_utils.py
