+++ diffsims/utils/atomic_diffraction_generator_support/generic_utils.py
